backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
motor==3.7.0
pymongo==4.11.3
python-socketio==5.12.1
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-dotenv==1.1.0
pydantic-settings==2.8.2
apscheduler==3.11.0
httpx==0.28.1
pytest==8.3.5
pytest-asyncio==0.25.3
anyio==4.9.0
mongomock-motor==0.0.34
